Added spot fx settle date
The CrossCcySwapEngine takes a spotFX input and the DiscountingCurrencySwapEngine takes an fxQuotes input.
These inputs assume the fx rate (or fx rates in the DiscountingCurrencySwapEngine case) settles without delay at the reference date.
But in the market there is always an fx spot settlement delay that equals - most of the time - 2 business days.
In order to accomodate for this market fact, I have added additional optional constructor arguments that define the number of fx settlement days and the corresponding calendar.
The corresponding similar adjustment was also done in the CrossCcyBasisSwapHelper constructor.
All these changes are backward compatible.